requirements.txt
pdfminer3
pyresparser
streamlit
pandas
pafy
plotly
pymysql
streamlit-tags
Pillow
youtube-dl
nltk
pdfminer3
spacy==2.3.5
requests
openai